Get started on making formatting error not a thing
